requirements.txt
llama-cloud>=0.1.0
llama-index-core>=0.12.0
openai>=1.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
